+++ package.json
@@ -19,6 +19,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-helmet-async": "1.3.0",
+    "react-loading": "2.0.3",
     "react-redux": "8.1.3",
     "react-router-dom": "6.16.0"
   },
