+++ src/components/favorite-list/index.ts
@@ -1,0 +1,1 @@
+export { FavoriteList } from './favorite-list';
