+++ src/components/footer/footer.tsx
@@ -1,10 +1,8 @@
 import { FC } from 'react';
 import { Logo } from '../logo';
 
-const Footer: FC = () => (
+export const Footer: FC = () => (
   <footer className="footer">
     <Logo logoType='footer' />
   </footer>
 );
-
-export default Footer;
