+++ src/pages/favorites-page.tsx
@@ -1,10 +1,11 @@
+import { FC } from 'react';
 import FavoriteEmpty from '../components/favorite-empty/favorite-empty';
-import FavoriteList from '../components/favorite-list/favorite-list';
-import Footer from '../components/footer/footer';
+import { FavoriteList } from '../components/favorite-list';
+import { Footer } from '../components/footer';
 import { MockFavorites } from '../mock/favorites';
 import { isEmpty } from '../utils';
 
-const Favorites = () => {
+export const FavoritesPage: FC = () => {
   const offers = MockFavorites;
   const isEmptyFavorites = isEmpty(offers);
   const cities = [...new Set(offers.map((offer) => offer.city.name))];
@@ -22,5 +23,3 @@ const Favorites = () => {
     </>
   );
 };
-
-export default Favorites;
